Revert "ZCS-8785:Fix for issue with data type range values for Mysql" (#1054)
This reverts commit 56a628bc6fc935f2876707e7645e3e0eee390a4d.

+++ store/src/java/com/zimbra/cs/store/external/ExternalMailboxBlob.java
@@ -30,7 +30,7 @@ import com.zimbra.cs.store.StoreManager;
  */
 public class ExternalMailboxBlob extends MailboxBlob {
 
-    protected ExternalMailboxBlob(Mailbox mbox, int itemId, long revision, String locator) {
+    protected ExternalMailboxBlob(Mailbox mbox, int itemId, int revision, String locator) {
         super(mbox, itemId, revision, locator);
     }
